feat: Create TripMapContent component
In anticipation of the v4 update of the react-leaflet package
